changing compliance kaldi to use on the fly generation

+++ test/common_utils/data_utils.py
@@ -68,7 +68,7 @@ def get_whitenoise(
     # so we only fork on CPU, generate values and move the data to the given device
     with torch.random.fork_rng([]):
         torch.random.manual_seed(seed)
-        tensor = torch.randn([sample_rate * duration], dtype=torch.float32, device='cpu')
+        tensor = torch.randn([int(sample_rate * duration)], dtype=torch.float32, device='cpu')
     tensor /= 2.0
     tensor *= scale_factor
     tensor.clamp_(-1.0, 1.0)
@@ -106,7 +106,7 @@ def get_sinusoid(
         dtype = getattr(torch, dtype)
     pie2 = 2 * 3.141592653589793
     end = pie2 * frequency * duration
-    theta = torch.linspace(0, end, sample_rate * duration, dtype=torch.float32, device=device)
+    theta = torch.linspace(0, end, int(sample_rate * duration), dtype=torch.float32, device=device)
     tensor = torch.sin(theta, out=None).repeat([n_channels, 1])
     tensor = convert_tensor_encoding(tensor, dtype)
     return tensor
